requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
pdf2image==1.14.0
Pillow==8.1.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytesseract==0.3.7
pyzbar==0.1.8
